requirements.txt
opencv-python
numpy
Pillow
requests
python-dotenv
python-sat
